feat(backend): Implement withdrawals and appreciation log retrieval
  - Implement retrieval of appreciation logs by day, week, month or
  quarter

+++ backend/app/server/cron/utils.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
-def calculateDailyAppreciation(amount: int, rate: int, duration: int) -> Decimal:
+def calculateDailyAppreciation(amount: int, rate: int, duration: int) -> int:
     duration_in_days = int(duration/86400)
-    return Decimal((rate * amount)/(100 * duration_in_days))
+    return (rate * amount)/(100 * duration_in_days)
